separate lita
