--- src/app/stay/page.module.scss
@@ -1,114 +1,0 @@
-@use "../../styles/mixin" as m;
-
-.Stay {
-  $padding-frame-vertical: 40px;
-  $padding-frame-horizontal: 48px;
-
-  width: 100%;
-  height: 100svh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: var(--color-stay-background);
-
-  & .Stay__texts {
-    // 準備中テキスト
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    text-align: center;
-    font-family: var(--font-family-text-jp);
-    color: var(--color-dark);
-
-    @include m.respondToSm {
-      padding: 20px;
-      font-size: 14px;
-    }
-
-    p {
-      line-height: 2.5;
-    }
-  }
-
-  & .Stay__igLink {
-    display: inline;
-    color: var(--color-text-gray);
-    text-decoration: underline;
-  }
-
-  & .Stay__reservationLink {
-    margin: 40px auto 0;
-    padding: 20px 24px;
-    display: flex;
-    justify-content: center;
-    align-items: baseline;
-    gap: 20px;
-    width: 360px;
-    font-size: 20px;
-    color: var(--color-dark);
-    font-family: var(--font-family-text-en);
-    border: 1px solid var(--color-dark);
-    border-radius: 40px;
-    // アニメーション設定
-    transition-duration: 0.4s;
-    transition-timing-function: ease-in-out;
-
-    @include m.respondToSm {
-      margin-top: 20px;
-      padding: 12px 24px;
-      width: 300px;
-      font-size: 17px;
-    }
-
-    @media (hover: hover) {
-      &:hover {
-        // color: var(--color-white);
-        border-color: var(--color-red);
-        background-color: var(--color-red);
-      }
-    }
-  }
-
-  & .Stay__linkArrow {
-    position: relative;
-    display: block;
-    width: 100px;
-    height: 0px;
-    border-top: 1px solid var(--color-black);
-    // アニメーション設定
-    transition-duration: 0.4s;
-    transition-timing-function: ease-in-out;
-
-    &::after {
-      content: "";
-      position: absolute;
-      top: -5px;
-      right: -3px;
-      display: block;
-      width: 12px;
-      height: 0px;
-      border-top: 1px solid var(--color-black);
-      transform: rotate(45deg);
-
-      // アニメーション設定
-      transition-duration: 0.4s;
-      transition-timing-function: ease-in-out;
-    }
-
-    @include m.respondToSm {
-      width: 60px;
-    }
-  }
-
-  & .Stay__legalNoticeLink {
-    margin: 40px auto 0;
-    display: block;
-    color: var(--color-text-gray);
-    text-align: center;
-    text-decoration: underline;
-
-    @include m.respondToSm {
-      font-size: 14px;
-    }
-  }
-}
