--- src/app/taste/page.module.scss
@@ -1,49 +1,0 @@
-@use "../../styles/mixin" as m;
-
-.Taste {
-  $padding-frame-vertical: 40px;
-  $padding-frame-horizontal: 48px;
-
-  width: 100%;
-  min-width: 1249px;
-  height: auto;
-
-  @include m.respondToSm {
-    width: 100%;
-    min-width: initial;
-  }
-
-  .Taste__noise {
-    position: fixed;
-    inset: 0;
-    z-index: var(--z-index-elevation-5);
-    pointer-events: none;
-    mix-blend-mode: soft-light;
-    opacity: 0.12;
-  }
-
-  .Taste__noiseGif {
-    position: absolute;
-    inset: 0;
-    background-image: url("/taste/noise.gif");
-    background-repeat: repeat;
-    background-size: 240px 240px;
-    animation: noiseJitter 1.6s steps(2, end) infinite;
-    will-change: background-position;
-    transform: translateZ(0);
-  }
-
-  .Taste__contents {
-    position: relative;
-    z-index: var(--z-index-elevation-1);
-    margin-top: -52px;
-    padding: 60px 0;
-    width: 100%;
-    background-color: var(--color-base-background);
-    border-radius: 60px;
-
-    @include m.respondToSm {
-      border-radius: 28px;
-    }
-  }
-}
