--- src/components/ContactForm/style.module.scss
@@ -1,166 +1,0 @@
-@use "../../styles/mixin" as m;
-
-.ContactForm {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 40px;
-
-  input,
-  textarea {
-    background-color: transparent;
-    border: none;
-    outline: none;
-  }
-
-  label {
-    display: flex;
-    align-items: center;
-  }
-
-  input[type="checkbox"] {
-    appearance: none;
-    -webkit-appearance: none;
-    -moz-appearance: none;
-
-    margin-right: 4px;
-    width: 14px;
-    height: 14px;
-    border: 1px solid #333;
-    border-radius: 2px;
-    background-color: white;
-    cursor: pointer;
-  }
-
-  input[type="checkbox"]:checked {
-    background-color: white;
-    position: relative;
-  }
-
-  input[type="checkbox"]:checked::after {
-    content: "✓"; /* チェックマークを疑似要素で */
-    position: absolute;
-    top: -3px;
-    left: -2px;
-    font-size: 16px;
-    color: black;
-  }
-
-  & .ContactForm__inputItem,
-  & .ContactForm__textareaItem {
-    width: 100%;
-  }
-
-  & .ContactForm__inputItemBody,
-  & .ContactForm__textarea {
-    margin-top: 8px;
-    padding: 12px 8px;
-    width: 100%;
-    border: 1px solid var(--color-black);
-    border-radius: 4px;
-    font-family: var(--font-family-text-jp);
-    color: var(--color-dark);
-
-    &::placeholder {
-      color: var(--color-stay-placeholder);
-    }
-  }
-
-  & .ContactForm__inputItemTitle {
-    @include m.respondToMd {
-      flex-shrink: 0;
-    }
-
-    &--required {
-      &::after {
-        content: " *";
-        color: var(--color-red);
-      }
-
-      @include m.respondToMd {
-        font-size: 13px;
-      }
-    }
-  }
-
-  & .ContactForm__inputItemTitleWithNotation {
-    display: flex;
-    align-items: flex-end;
-    gap: 4px;
-
-    & .ContactForm__inputItemNotation {
-      font-size: 13px;
-      color: var(--color-text-gray);
-
-      @include m.respondToMd {
-        font-size: 11px;
-      }
-    }
-  }
-
-  & .ContactForm__textarea {
-    height: 200px;
-  }
-
-  & .ContactForm__confirm {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    width: 100%;
-
-    p {
-      font-size: 14px;
-      line-height: 1.6;
-
-      @include m.respondToMd {
-        font-size: 13px;
-      }
-    }
-
-    label {
-      font-size: 15px;
-
-      @include m.respondToMd {
-        font-size: 14px;
-      }
-    }
-  }
-
-  & .ContactForm__accommodationLink {
-    text-decoration: underline;
-  }
-
-  & .ContactForm__inputItemBody--error {
-    border: 1px solid var(--color-red);
-    background-color: var(--color-cream);
-  }
-
-  & .ContactForm__errorMessage {
-    color: var(--color-red);
-    font-size: 0.875rem;
-    margin-top: 4px;
-  }
-
-  & .ContactForm__submitButton {
-    margin-top: 16px;
-    padding: 16px 0;
-    width: 100%;
-    background-color: var(--color-red);
-    color: var(--color-white);
-    font-size: 16px;
-    font-weight: bold;
-    border-radius: 4px;
-    transition-duration: 0.2s;
-    transition-timing-function: ease-in;
-
-    &:hover {
-      opacity: 0.7;
-    }
-  }
-
-  & .ContactForm__submitStatus {
-    margin-top: -20px;
-    color: var(--color-red);
-    text-decoration: underline;
-  }
-}
