--- src/components/SlideTimer/style.module.scss
@@ -1,53 +1,0 @@
-.Timer {
-  position: absolute;
-  right: 24px;
-  bottom: 24px;
-  width: var(--timer-size);
-  height: var(--timer-size);
-  pointer-events: none;
-  z-index: 3;
-}
-
-.Timer__svg {
-  transform: rotate(-90deg);
-} /* 12時起点 */
-
-.Timer__track {
-  fill: none;
-  stroke: rgba(255, 255, 255, 0.35);
-}
-
-.Timer__progress {
-  fill: none;
-  stroke: var(--color-white); /* 実際の色はインラインで上書き可 */
-  stroke-linecap: round;
-
-  /* dasharray / dashoffset はインライン指定（Safari 安定）*/
-  will-change: stroke-dashoffset;
-  filter: drop-shadow(0 0 2px rgba(0, 0, 0, 0.2));
-
-  /* keyframes 名は CSS Module 内で完結させる（ハッシュ対策） */
-  animation-name: timerProgress;
-  animation-timing-function: linear;
-  animation-fill-mode: forwards;
-  animation-duration: var(--duration, 5500ms);
-}
-
-@keyframes timerProgress {
-  to {
-    stroke-dashoffset: 0;
-  }
-}
-
-/* 暗い背景向けのトラック色 */
-.Timer[data-theme="dark"] .Timer__track {
-  stroke: rgba(0, 0, 0, 0.35);
-}
-
-/* 低モーション環境を尊重（propsで無効化可） */
-@media (prefers-reduced-motion: reduce) {
-  .Timer[data-respect-reduced="true"] .Timer__progress {
-    animation: none !important;
-    stroke-dashoffset: 0 !important;
-  }
-}
