+++ src/styles/globals.scss
@@ -1,53 +1,1 @@
 @import "./_reset.scss";
-
-// body {
-//   // font-family: mr-eaves-sans, hina-mincho, sans-serif;
-//   font-family: ltc-nicholas-cochin-pro, hina-mincho, sans-serif;
-//   font-weight: 400;
-//   color: var(--color-black);
-// }
-
-// :root {
-//   // basic color
-//   --color-black: #202020;
-//   --color-white: #fff;
-//   --color-gold: #b6a66c;
-//   --color-cream: #f0efea;
-//   --color-green: #215256;
-//   --color-blue: #263d5f;
-//   --color-red: #7b342e;
-//   --color-brown: #795542;
-//   --color-base-background: var(--color-white);
-
-//   // ---- taste color ----
-//   // text color
-//   --color-text-black: var(--color-black);
-//   --color-text-gray: #616161;
-//   --color-text-gray-2: #d8d8d8;
-//   --color-text-gold: var(--color-gold);
-//   // border color
-//   --color-border-gray: rgba(0, 0, 0, 0.07);
-
-//   // ---- stay color ----
-//   --color-stay-text-black: var(--color-black);
-//   --color-stay-text-white: var(--color-white);
-//   --color-stay-background: var(--color-cream);
-//   --color-stay-placeholder: #a6a6a6;
-
-//   // font-family
-//   --font-family-title: ltc-nicholas-cochin-pro, hina-mincho, sans-serif;
-//   --font-family-gothic: mr-eaves-sans, hina-mincho, sans-serif; /** ほとんど使わない */
-//   --font-family-text-jp: hina-mincho, sans-serif;
-//   --font-family-text-en: hina-mincho, sans-serif;
-
-//   // letter-spacing
-//   --letter-space-text-jp: 0.02em;
-//   --letter-space-text-en: 0.01em;
-
-//   // z-index
-//   --z-index-elevation-1: 100000;
-//   --z-index-elevation-2: 300000;
-//   --z-index-elevation-3: 600000;
-//   --z-index-elevation-4: 900000;
-//   --z-index-elevation-5: 2147483647;
-// }
